+++ src/gs/game/actor/Actor.cpp
@@ -113,7 +113,7 @@ void Actor::setPosition(s32 const x, s32 const y, s32 const z) { setPosition(Pos
 void Actor::setPosX       (s32 const x) { component<Position>()->x           = x; }
 void Actor::setPosY       (s32 const y) { component<Position>()->y           = y; }
 void Actor::setPosZ       (s32 const z) { component<Position>()->z           = z; }
-void Actor::setOrientation(s32 const o) { component<Position>()->orientation = o; }
+void Actor::setOrientation(u16 const o) { component<Position>()->orientation = o; }
 
 void Actor::setTeam(Team const team) { _impl->team = team; }