+++ src/gs/game/components/Position.hpp
@@ -12,9 +12,13 @@ struct Position : public Component
 {
     DECLARE_PACKET_SERIALIZATION_OPERATORS(Position);
 
-    s32 x, y, z, orientation = 0;
+    s32 x, y, z;
+    u16 orientation;
 
-    explicit Position(s32 const x_ = 0, s32 const y_ = 0, s32 const z_ = 0) noexcept
-        : x(x_), y(y_), z(z_)
+    explicit Position(s32 const x_ = 0, s32 const y_ = 0, s32 const z_ = 0, u16 const orientation_ = 0) noexcept
+        : x{x_}
+        , y{y_}
+        , z{z_}
+        , orientation{orientation_}
     {}
 };