+++ src/gs/game/directories/ProfessionDirectory.hpp
@@ -21,12 +21,14 @@ class ProfessionDirectory
     struct ProfessionInfo
     {
         std::string name;
-        float       maxHp, hpFlatPerLevel, hpMultiplierPerLevel;
-        float       maxMp, mpFlatPerLevel, mpMultiplierPerLevel;
-        float       maxCp, cpFlatPerLevel, cpMultiplierPerLevel;
-        Profession  profession, parentProfession;
-        u8          minimumLevel;
-        bool        canBeSubclassed;
+        float       maxHp{}, hpFlatPerLevel{}, hpMultiplierPerLevel{};
+        float       maxMp{}, mpFlatPerLevel{}, mpMultiplierPerLevel{};
+        float       maxCp{}, cpFlatPerLevel{}, cpMultiplierPerLevel{};
+
+        std::optional<Profession> parentProfession;
+        Profession                profession{};
+        u8                        minimumLevel{1};
+        bool                      canBeSubclassed{};
 
         void applyBaseStats(Actor &) const;
     };
@@ -36,11 +38,14 @@ public:
 
 public:
     static auto count() -> size_t;
-    static auto find(Profession) -> OptRef<ProfessionInfo const>;
+    static auto find              (Profession) -> OptRef<ProfessionInfo const>;
+    static auto parent            (Profession) -> OptRef<ProfessionInfo const>;
+    static auto startingProfession(Profession) -> OptRef<ProfessionInfo const>;
+    static auto rank              (Profession) -> u8;
 
 public:
     static void load();
 
 private:
-    static std::unordered_map<u32, ProfessionInfo> _professions;
+    static std::unordered_map<Profession, ProfessionInfo> _professions;
 };